+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 *The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)*
 
+## 0.2.0 - UNRELEASED
+
+### Added
+
+- Add all data types without parameters for PostgreSQL. 
+
 ## 0.1.3 - 2023-10-26
 
 ### Fixed
@@ -20,7 +26,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
-- Add model name to the auto-generated migration name
+- Add model name to the auto-generated migration name.
 
 ## 0.1.1 - 2023-09-25